+++ rotkehlchen/data/importer.py
@@ -4,15 +4,29 @@ import logging
 from collections import defaultdict
 from itertools import count
 from pathlib import Path
-from typing import Any, Dict, List, Tuple
+from typing import Any, Dict, List, Optional, Tuple
 
 from pysqlcipher3 import dbapi2 as sqlcipher
 
 from rotkehlchen.accounting.ledger_actions import LedgerAction, LedgerActionType
-from rotkehlchen.assets.converters import asset_from_cryptocom, asset_from_nexo, asset_from_uphold
+from rotkehlchen.assets.converters import (
+    asset_from_binance,
+    asset_from_cryptocom,
+    asset_from_nexo,
+    asset_from_uphold,
+)
 from rotkehlchen.assets.utils import symbol_to_asset_or_token
 from rotkehlchen.constants.assets import A_BSQ, A_BTC, A_DAI, A_SAI, A_USD
 from rotkehlchen.constants.misc import ZERO
+from rotkehlchen.data.binance_import_utils import (
+    BinanceCsvRow,
+    BinanceDepositWithdrawEntry,
+    BinanceEntry,
+    BinancePOSEntry,
+    BinanceSingleEntry,
+    BinanceStakingRewardsEntry,
+    BinanceTradeEntry,
+)
 from rotkehlchen.db.dbhandler import DBHandler
 from rotkehlchen.db.ledger_actions import DBLedgerActions
 from rotkehlchen.errors.asset import UnknownAsset
@@ -33,6 +47,18 @@ logger = logging.getLogger(__name__)
 log = RotkehlchenLogsAdapter(logger)
 
 SAI_TIMESTAMP = 1574035200
+
+SINGLE_BINANCE_ENTRIES = [
+    BinanceDepositWithdrawEntry(),
+    BinanceStakingRewardsEntry(),
+    BinancePOSEntry(),
+]
+
+MULTIPLE_BINANCE_ENTRIES = [
+    BinanceTradeEntry(),
+]
+
+BINANCE_TRADE_OPERATIONS = {'Buy', 'Sell', 'Fee'}
 
 
 def remap_header(fieldnames: List[str]) -> List[str]:
@@ -1433,4 +1459,136 @@ Activity from uphold with uphold transaction id:
                     continue
                 except KeyError as e:
                     return False, str(e)
+        return True, ''
+
+    @staticmethod
+    def _group_binance_rows(
+        rows: List[BinanceCsvRow],
+        **kwargs: Dict[str, Any],
+    ) -> Tuple[int, Dict[Timestamp, List[BinanceCsvRow]]]:
+        """Groups Binance rows by timestamp and deletes unused columns"""
+        multirows: Dict[Timestamp, List[BinanceCsvRow]] = defaultdict(list)
+        skipped_count = 0
+        formatstr = kwargs.get('timestamp_format')
+        for csv_row in rows:
+            try:
+                timestamp = deserialize_timestamp_from_date(
+                    date=csv_row['UTC_Time'],
+                    formatstr=formatstr if formatstr is not None else '%Y-%m-%d %H:%M:%S',  # type: ignore  # noqa: E501
+                    location='binance',
+                )
+                csv_row['Coin'] = asset_from_binance(csv_row['Coin'])
+                csv_row['Change'] = deserialize_asset_amount(csv_row['Change'])
+                csv_row.pop('UTC_Time', None)
+                csv_row.pop('User_ID', None)
+                csv_row.pop('Account', None)
+                csv_row.pop('Remark', None)
+                multirows[timestamp].append(csv_row)
+            except (DeserializationError, UnknownAsset) as e:
+                log.warning(f'Skipped binance csv row {csv_row} because of {str(e)}')
+                skipped_count += 1
+            except KeyError as e:
+                log.error(f'Malformed binance csv columns! Broke on row {csv_row}. {str(e)}')
+                return len(rows), {}
+
+        return skipped_count, multirows
+
+    def _process_single_binance_entries(
+        self,
+        timestamp: Timestamp,
+        rows: List[BinanceCsvRow],
+    ) -> Tuple[Dict[BinanceSingleEntry, int], List[BinanceCsvRow]]:
+        """Processes binance entries that are represented with a single row in a csv file"""
+        processed: Dict[BinanceSingleEntry, int] = defaultdict(int)
+        ignored: List[BinanceCsvRow] = []
+        for row in rows:
+            for single_entry_class in SINGLE_BINANCE_ENTRIES:
+                if single_entry_class.is_entry(row['Operation']):
+                    single_entry_class.process_entry(
+                        db=self.db,
+                        db_ledger=self.db_ledger,
+                        timestamp=timestamp,
+                        data=row,
+                    )
+                    processed[single_entry_class] += 1
+                    break
+            else:
+                ignored.append(row)
+        return processed, ignored
+
+    def _process_multiple_binance_entries(
+        self,
+        timestamp: Timestamp,
+        rows: List[BinanceCsvRow],
+    ) -> Tuple[Optional[BinanceEntry], int]:
+        """Processes binance entries that are represented with 2+ rows in a csv file"""
+        for multiple_entry_class in MULTIPLE_BINANCE_ENTRIES:
+            if multiple_entry_class.are_entries([row['Operation'] for row in rows]):
+                processed_count = multiple_entry_class.process_entries(
+                    db=self.db,
+                    timestamp=timestamp,
+                    data=rows,
+                )
+                return multiple_entry_class, processed_count
+        return None, 0
+
+    def _process_binance_rows(self, multi: Dict[Timestamp, List[BinanceCsvRow]]) -> None:
+        stats: Dict[BinanceEntry, int] = defaultdict(int)
+        skipped_rows: List[Any] = []
+        for timestamp, rows in multi.items():
+            single_processed, rows_without_single = self._process_single_binance_entries(
+                timestamp=timestamp,
+                rows=rows,
+            )
+            for entry_type, amount in single_processed.items():
+                stats[entry_type] += amount
+
+            multiple_type, multiple_count = self._process_multiple_binance_entries(
+                timestamp=timestamp,
+                rows=rows_without_single,
+            )
+            if multiple_type is not None and multiple_count > 0:
+                stats[multiple_type] += multiple_count
+                rows_without_multiple = []
+            else:
+                rows_without_multiple = rows_without_single
+
+            if len(rows_without_multiple) > 0:
+                skipped_rows.append([timestamp, rows_without_multiple])
+
+        skipped_nontrade_rows = []
+        skipped_trade_rows = []
+        for skipped_rows_group in skipped_rows:
+            cur_operations = {el['Operation'] for el in skipped_rows_group[1]}
+            if cur_operations.issubset(BINANCE_TRADE_OPERATIONS):
+                skipped_trade_rows.append(skipped_rows_group)
+            else:
+                skipped_nontrade_rows.append(skipped_trade_rows)
+        total_found = sum(stats.values())
+        skipped_count = 0
+        for _, rows in skipped_rows:
+            skipped_count += len(rows)
+        log.debug(f'Skipped Binance trade rows: {skipped_trade_rows}')
+        log.debug(f'Skipped Binance non-trade rows {skipped_nontrade_rows}')
+        log.debug(f'Total found Binance entries: {total_found}')
+        log.debug(f'Total skipped Binance csv rows: {skipped_count}')
+        log.debug('Binance import stats: {}'.format(
+            [{type(entry_class).__name__: amount} for entry_class, amount in stats.items()],
+        ))
+        if skipped_count > 0:
+            self.db.msg_aggregator.add_warning(
+                f'Skipped {skipped_count} rows during processing binance csv file. '
+                f'Check logs for details',
+            )
+
+    def import_binance_csv(self, filepath: Path, **kwargs: Any) -> Tuple[bool, str]:
+        with open(filepath, 'r', encoding='utf-8-sig') as csvfile:
+            input_rows = list(csv.DictReader(csvfile))
+            skipped_count, multirows = self._group_binance_rows(rows=input_rows, **kwargs)
+            if skipped_count > 0:
+                self.db.msg_aggregator.add_warning(
+                    f'{skipped_count} Binance rows have bad format. Check logs for details.',
+                )
+            self._process_binance_rows(multirows)
+
         return True, ''